An alternative dump if we need it
It may be that sling.io, when used via Python, tries to read whole
tables into RAM and construct the CSV before streaming to S3.

This operates differently:

1. Each table is
2. Written to disk as CSV
3. By chunked SQL reads in pandas and then
4. Zipped

This approach has been used in other places in our codebase to get
around our large filesizes/query sizes. It should be the case that this
will run in less RAM than the sling.io approach.

This has only been tested locally so far.