Results from walking the code for `audit`
We wanted to make sure all pathways for the audit submission do not
reference the `audit` table in a way that might break submissions at
this time. (We found a small bug last week, and a small bug this week.)
We think all the pathways are clean.

* Some consistency changes were made for logging
* Some strings were replaced with constants
* We nested _compare_errors() under an `if` in most cases

The last point should make our logs quieter. We were comparing
validation errors on the `sac` and `sot` in ALL cases, when in truth, we
should only do the comparison in cases where there is an `audit` record
in the first place. We have hundreds of audits for which we only have a
SAC, and this was causing a lot of noise in the logs.