Making logs quieter
There are two `warnings` and `info` that do not need to be in our logs.
We added comments in both cases. Because the code involved is 18 months
and 22 months in production, it is clear that we can turn these to
debugs, because the behavior they are logging is, at this point,
expected and handled behavior.