Performance enhancements
Previously, this migration script would gradually slow down the further along the progress. And we suspect this to be the problematic line:
`queryset = queryset.exclude(report_id__in=Audit.objects.all().values("report_id"))`.

Every single batch of 100 records would re-run this, and the long story short of it is that the `.exclude(report_id__in=` syntax will take DRAMATICALLY longer the bigger the `Audit` table.
We also suspect `while queryset.count() == 0` to not provide us any favors, so there is some refactoring here to account for the removal of this line. We do not need to group the data in batches outside of logging information to monitor performance.