Pushing improved filtering
Per slack comment:

I'm pushing a version of the comparison script that changes how the ignore file is structured. I've included ignore2.json .

This isn't fast/great. But, it now encodes 12K audit report IDs that have incorrect dates for "ready_for_certification_date," and correctly skips only those reports based on that field. (It doesn't skip the report wholesale; it skips just that field in those reports). So, it is a minimal approach to skipping.

It wants more testing, and we'll have to expand this to include the actual values from production at the time that we run it in production. (However... once we do this in staging, the list shouldn't change... because it should only be old audits that have this issue, not new ones.)