1.1.0 vs 1.2.0 API check improvements (#5077)
* Metabase local stack

* typo

* Changes to allow ignoring columns

There are some columns that have historical values that we need to
ignore in comparing the API.

This is a start on filtering this better.

It's not 100%, but it lets us do a test pass against all of the data in
API v1.1 and v1.2.

* Chasing down more oddities

We're zeroing in on having the data/comparisons right.

We think the timezone needs to change, or we need to drop it (use UTC)
and consider this a data improvement in v1.2.

* Adding the ability to skip `stuck` audits

* Pushing improved filtering

Per slack comment:

I'm pushing a version of the comparison script that changes how the ignore file is structured. I've included ignore2.json .

This isn't fast/great. But, it now encodes 12K audit report IDs that have incorrect dates for "ready_for_certification_date," and correctly skips only those reports based on that field. (It doesn't skip the report wholesale; it skips just that field in those reports). So, it is a minimal approach to skipping.

It wants more testing, and we'll have to expand this to include the actual values from production at the time that we run it in production. (However... once we do this in staging, the list shouldn't change... because it should only be old audits that have this issue, not new ones.)

* Interim push to handle pagination, more ignores

Focusing in on dictionary comparison. Now handling more
objects/pagination

* Incremental. Fixing skip logic.

The skip logic still looks like it can be improved when doing a
compare_any_order. It feels like the slowest part of the whole thing.

I'm wondering if a pre-filter of objects would be better, and then
process the data after filtering. :/

* Incremental

This improves performance in a number of ways.

* Adding incremental

Minor changes for performance

* Add an `audit_year` option to the API comparison scripts

Intended for smaller tables without acceptance dates.

* Default timing_window_size to 1 to avoid divide by zero with small batches.

* Lint

* db3 merge fix

* Lint

* Creating FacTest to appease linter; adding timeout

* Lint

* Lint

* Lint

* Lint

* Updating ignore list

* Adding notes_to_sefa handling

* Err message when audit_year not supplied

* Fixing sefa result

---------

Co-authored-by: Alex Steel <130377221+asteel-gsa@users.noreply.github.com>
Co-authored-by: Matt Jadud <matthew.jadud@gsa.gov>
Co-authored-by: James Person <jperson1@umbc.edu>