Adds "redisseminate()" and moves closer to complete
This now will relplace UEI or EIN records, and adds `redisseminate()` to
the SAC, because we'll need it. (And, it is safer to put there than
here.)

Some more work to do, but I wanted to get it pushed in an interim state.

Many thanks to Phil for working through the logic in `.save()`.