Update UEIs or EINs for submitted audits (#5108)
* First steps on UEI update

This is a start. I need to refresh my local before going further.

1. Entries validated.
2. Changes made.
3. Save implemented.

* Adds "redisseminate()" and moves closer to complete

This now will relplace UEI or EIN records, and adds `redisseminate()` to
the SAC, because we'll need it. (And, it is safer to put there than
here.)

Some more work to do, but I wanted to get it pushed in an interim state.

Many thanks to Phil for working through the logic in `.save()`.

* Adding hashing, __eq__

In order to add things to set()s, I needed to add a __hash__ function to
the model. I hash the string representation. (This will not impact
anything in production.)

* Updating tool, adding tests

This adds some unit tests, and also breaks things apart (mostly for
testing).

* More tests, validations on inputs

This now defends nicely against bad inputs, and only works if the audit
is already dissemianted.

* Adding in code for supporting redissemination

Adding a DISSEMINATED->DISSEMINATED transition, so that we can extend
the transition_name and transition_date fields to include additional
dissemination notices when we... redisseminate a record.

* Final changes before merge.

* Updating throws, linting

`make lint` was run inside the container.

Lets see if that aligns with what we do in the cloud.

* Moving atomic()

Making sure that the curation tracking happens atomically.

* Adding in docs

We'll want to update this for other commands. I don't think we have any
notes anywhere about our management commands.

* The ordering of curation tracking matters

We cannot

atomic.transaction():
  with curation_tracking():
    ...

Why? Because curation tracking creates AND destroys triggers.

As a result, doing that under a transaction causes the process/DB to
lock.

Instead, we MUST

with curation_tracking():
  atomic.transaction():
    ...

so that we enable the triggers, do the atomic work, and disable the
triggers for tracking changes.

* A new migration comes with this work

Why?

Becuase we are adding to the choices for a submission event. Therefore,
those choices are reflected in the model.

* Fixing .save()

When updating the .save() in the SAC, we missed that we should ALWAYS
create the submission event. Our unit tests saved us.

All tests pass, comments improved the code, and we think we left it in a
better state. HT to JamesP and PhilD.

* Updates based on PD's suggestions

Some renaming, a few comments removed, table name fix.

* Quick linting fix for the review

---------

Co-authored-by: Phil Dominguez <142051477+phildominguez-gsa@users.noreply.github.com>
Co-authored-by: James Person <jperson1@umbc.edu>