Findings Summary Report (#5140)
* Adding a "findings summary view"

We've been asked for something like this forever.

We've had the code for over a year.

This is a "v1." Perhaps we only generate it for Feds.

Pushing it as a "what if," so we can discuss it.

* Continued exploration of improved report output

This has been requested for a long time. We had most of the logic.

Performant enough with a few thousand results.

* Some fixes. Extra protections, some tests.

* Fixing linting, werd try/catch

I believe, based on unit testing and manual testing, that this will
protect the code in question. These are both cases where I almost want a
bare `except: pass`, but I've chosen not to.

We'll need to keep our eyes on this, but it looks good.

---------

Co-authored-by: Matt Jadud <matthew.jadud@gsa.gov>