Improved menubash - tooling for working with data locally (#5162)
* Incremental

This has a working end-to-end for the preparation script.

Next is the script for devs using the data.

* Working prep and load

Going to do another test and documentation, but this looks like
end-to-end test data gen and prep.

* Docs, minor changes

This gives us docs for the dev-facing tool.

Should be good for other people to use/tweak/find improvements.

* Adding a link to the santizied dump folder

Link to GDrive added

* Improved handling/checking for things

This makes sure the service key is present, and creates it if not.

Also checks for apps.

* Improved CF handling

* Updating command in docs

* Removing `readlink`

Does not work on M1/M2 Macs.

* More prompts

As long as my colleagues are happy. :D

* Adds a positive confirmation of prod

As long as we're checking, we should say what we see.

* Mistake w/ v15/v17 confusion yesterday.

* Updated to use Dockerfiles

This allowed the scripts to work back-to-back. More testing needed.

* Should... work?

Everything now runs in the app container, which eliminates some
problems, creates others, but should eliminate issues of things working
differently on different machines.

* Removing unneeded files, counts

Removes table counts (no longer meaningful) and files not needed
anymore with the pair of menus.

* Adding a positive confirmation

One action was kinda quiet.

* Switching container base

We were using

python:3.12-slim

which I *think* became a Trixie (Debian 13) image very queitly as of
8/13.

Changing it to

3.12.11-slim-bookworm

to see if that locks it to Deb 12.