Adding tribal resubmission release banner (#5174)
* Adding tribal resubmission release banner

* Making separate test file for banners

* Adding back valid report_id case

* Typos

* Forgot banner file

* Update to previous_is_public logic

This updates the context to look one audit back to get the
`is_tribal_information_...public` value. It was grabbing the value from
the current sac, which would give incorrect behavior, I think.

* Adding a comment, looks good.

We walked through the code, considering the not-tribal -> tribal
resubmission case, and added a comment for future us.

LGTM.

---------

Co-authored-by: Matt Jadud <matthew.jadud@gsa.gov>