Jadudm/resubmission analysis (#5161)
* Analyzing resubmissions

In order to implement resubmission, we'll have to curate the existing
database.

That implies we know which audits were "resubmissions" of each-other in
the past.

This is a clustering algorithm that introduces an "audit_distance" to
determine if audits are similar enough to be considered resubmissions of
each-other.

* Use deepcopy

* Improves ordering of records

Orders the queryset of records after the query; the in-SQL/model
ordering was not useful.

* Cleanup/breakup of code, tests

Added unit tests. More are needed, but the core is starting to be
covered.

Broke the library up into sensible parts (for
clarity/documentation/testing), and so we could use an SQLite pull from
the API. This provided two pathways to validation: the Postgres data and
the SQLite data.

* Updating with linking code.

* Generate a CSV, too.

We will want that for curation publication, along with the .md.

* Linting

* Removing unneeded file.

The SQLite stuff was from testing before it became tied into the
containerized stack.

* Another SQLite file

Not needed anymore.

* Updating comments

* Missed one of the merge conflicst.

* The merge mangled the paren groupings.

* Linting, test fixes

In particular, under unit tests, the triggers could be created more than
once. This makes them `CREATE OR REPLACE`, which is not ideal, but it is
a safe way to avoid this problem (locally and in production).

* Update backend/curation/curationlib/export_resubmission_clusters.py

LGTM. TY.

Co-authored-by: Phil Dominguez <142051477+phildominguez-gsa@users.noreply.github.com>

* Update backend/curation/curationlib/export_resubmission_clusters.py

LGTM. TY.

Co-authored-by: Phil Dominguez <142051477+phildominguez-gsa@users.noreply.github.com>

* Update backend/curation/curationlib/export_resubmission_clusters.py

LGTM. TY.

Co-authored-by: Phil Dominguez <142051477+phildominguez-gsa@users.noreply.github.com>

* Update backend/curation/curationlib/export_resubmission_clusters.py

LGTM. TY.

Co-authored-by: Phil Dominguez <142051477+phildominguez-gsa@users.noreply.github.com>

* Adding comment

* Fixing generation of requirements, merge

---------

Co-authored-by: Phil Dominguez <142051477+phildominguez-gsa@users.noreply.github.com>