Adding two lines, filtering tags
This adds one condition to tags:

If you are an authenticated user with Tribal data access, then the tags
are added. Otherwise, they are not.

This is the same as saying "You are a Federal user who has privileged
data access." Those are the only users who should see the tags.

Confirmed by viewing results:

* Unauthenticated
* Authenticated
* Authenticated with `read-tribal`