Feat(ui): ] Add resubmitted and most recent tags to the search result… (#5193)
* Feat(ui): ] Add resubmitted and most recent tags to the search results UI
#5119

* Generate fake dissem_general resub data

This supports the inserting of tags in the interface.

* tags showing in basic search

* Merge branch 'main' into resub_search_tag

* tags updated and working for advance/basic

* black .

* Delete backend/dev-requirements.txt

* linting

* fixed linting

* black .

* added updates, fixed tests, reformatted using black

* Update requirements.txt

* removed any for keys, replaced with str

* Adding two lines, filtering tags

This adds one condition to tags:

If you are an authenticated user with Tribal data access, then the tags
are added. Otherwise, they are not.

This is the same as saying "You are a Federal user who has privileged
data access." Those are the only users who should see the tags.

Confirmed by viewing results:

* Unauthenticated
* Authenticated
* Authenticated with `read-tribal`

* fixed wrapping and bold

* reverted

---------

Co-authored-by: Matt Jadud <matthew.jadud@gsa.gov>