Simpler CSV export
This bash script replaces the use of Sling for CSV export.

Why?

1. Sling introduces upstream security concerns that we must constantly
   respond to.
2. The sling python library does not import well in our multi-platform
   (Mac, Linux, Windows) Docker stack.
3. The procedss is *slow*. Many hours?

This runs in 15m locally; ymmv in the cloud.

It takes a list of endpoints from an API, and does the following query
against each:

  local query="\COPY "
  query+="( SELECT * FROM ${API_VERSION}.${endpoint} "
  query+="  WHERE report_id in ( "
  query+="    SELECT report_id from ${API_VERSION}.general "
  query+="    WHERE fac_accepted_date >= '${start_date}' "
  query+="    AND fac_accepted_date <= '${end_date}' "
  query+=")) "
  query+="TO '${ROOT}/${endpoint}.csv' "
  query+="WITH (FORMAT CSV, HEADER, DELIMITER ',');"

This creates a CSV file in the filesystem that we then use the aws-cli
to ship over to S3. By putting the file in the correct place, the static
site can pick up the CSVs.

Initial checks suggest this is exporting data in a manner consistent
with the previous tooling.

TODO:

- [ ] Test in `staging`
- [ ] Remove the sling tooling
- [ ] Modify the GH action to use this