Updating the view
This updates the view to accept only one audit. It also implements the
access controls as discussed, with minor changes.

It is unclear whether the diff should *always* be available, even after
submission. In this way, an oversight official can refer to the diff
page, and the owners/participants on the audit will still be able to see
it. It will continue to prevent people *not* associated with the audit
from seeing the diff, so that seems OK.