It was in the test db
The issue was with how the test DB was being constructed. There were
invariants being broken (?!), and as a result, I had reports with
duplicate report IDs. This broke selection logic in the code.

Once all of that was figured out, the view tests were fine. The issue
was entirely with test setup/data construction.